[refactor] Swipe action & save bug fixed
